requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
contourpy==1.3.1
crispy-bootstrap4==2024.10
cycler==0.12.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.16
django-allauth==65.3.0
django-cloudinary-storage==0.3.0
django-cloudistatic==0.0.1
django-crispy-forms==2.3
fonttools==4.55.1
gunicorn==20.1.0
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.0
sqlparse==0.5.2
urllib3==1.26.20
whitenoise==5.3.0
